fix(FIN-038): move filter and schema to separate domain folder
