+++ src/app/api/tracking-operation/update/[id]/route.ts
@@ -1,5 +1,5 @@
 import { createRoute } from '@backend/shared/utils/create-route.util';
-import { TrackingOperationSchema } from '@common/domains/lookups/schemas/tracking-operation.schema';
+import { TrackingOperationSchema } from '@common/domains/tracking-operation/schema/tracking-operation.schema';
 import { GetUserIdTransformer } from '@backend/shared/transformers/get-user-id.transformer';
 import { GetIntegerParamPipe } from '@backend/shared/pipes/get-integer-param.pipe';
 import { AuthGuard } from '@backend/entities/user/infrastructure/auth.guard';
