+++ src/common/domains/tracking-operation/schema/tracking-operation.schema.ts
