+++ src/server/entities/tracking-operation/domain/tracking-operation-api.filter.ts
@@ -1,4 +1,4 @@
-import { type TrackingOperationFilter } from '@common/domains/lookups/filters/tracking-operation.filter';
+import { type TrackingOperationFilter } from '@common/domains/tracking-operation/filter/tracking-operation.filter';
 
 export interface TrackingOperationApiFilter extends TrackingOperationFilter {
   userId: number;
